+++ celescope/capture_virus/__init__.py
@@ -2,12 +2,12 @@ __STEPS__ = [
     'sample',
     'barcode',
     'cutadapt',
-    "STAR_virus",
+    "star_virus",
     "count_capture_virus",
     "analysis_capture_virus",
 ]
 __ASSAY__ = 'capture_virus'
 
 IMPORT_DICT = {
-    'STAR_virus': 'celescope.rna_virus',
+    'star_virus': 'celescope.rna_virus',
 }