+++ celescope/capture_virus/multi_capture_virus.py
@@ -4,20 +4,20 @@ from celescope.tools.multi import Multi
 
 class Multi_capture_virus(Multi):
 
-    def STAR_virus(self, sample):
-        step = 'STAR_virus'
+    def star_virus(self, sample):
+        step = 'star_virus'
         cmd_line = self.get_cmd_line(step, sample)
-        input_read = f'{self.outdir_dic[sample]["cutadapt"]}/{sample}_clean_2.fq{self.fq_suffix}'
+        fq = f'{self.outdir_dic[sample]["cutadapt"]}/{sample}_clean_2.fq{self.fq_suffix}'
         cmd = (
             f'{cmd_line} '
-            f'--input_read {input_read} '
+            f'--fq {fq} '
         )
         self.process_cmd(cmd, step, sample, m=self.args.starMem, x=self.args.thread)
 
     def count_capture_virus(self, sample):
         step = 'count_capture_virus'
         cmd_line = self.get_cmd_line(step, sample)
-        virus_bam = f'{self.outdir_dic[sample]["STAR_virus"]}/{sample}_virus_Aligned.sortedByCoord.out.bam'
+        virus_bam = f'{self.outdir_dic[sample]["star_virus"]}/{sample}_virus_Aligned.sortedByCoord.out.bam'
         cmd = (
             f'{cmd_line} '
             f'--virus_bam {virus_bam} '